requirements.txt
Flask==3.0.0
supabase==2.15.0
python-dotenv==1.0.0